refactor : 푼 문제 리스트도 반환 (#187)

+++ src/main/java/org/ezcode/codetest/domain/submission/dto/DailyCorrectCount.java
@@ -5,6 +5,7 @@ import java.util.Set;
 
 public record DailyCorrectCount(
     LocalDate date,
-    Long count,
+    int count,
     Set<Long> problemIds
-) {  }
+) {
+}
